Requirements
I've also added the commands needed to install the package that contains javeria's dataset. In the confirmation email you can write something like: "To ensure that you benefit from the class please make sure you have the required software mentioned <here> installed on your laptops so that you can follow the drills along with our mentors." and give them a link to this page. They shouldn't fail to do this if they're worth it.